+++ browser-switch/src/main/java/com/braintreepayments/browserswitch/BrowserSwitchActivity.java
@@ -1,8 +1,10 @@
 package com.braintreepayments.browserswitch;
 
 import android.app.Activity;
+import android.content.Intent;
 import android.os.Bundle;
 
+import androidx.annotation.NonNull;
 import androidx.annotation.VisibleForTesting;
 
 /**
@@ -11,6 +13,8 @@ import androidx.annotation.VisibleForTesting;
  * finishes during {@link Activity#onCreate(Bundle)}.
  */
 public class BrowserSwitchActivity extends Activity {
+
+    private static Intent sReturnIntent;
 
     @VisibleForTesting
     // TODO: Refactor browser-switch-android to allow injection of a custom url scheme
@@ -21,6 +25,19 @@ public class BrowserSwitchActivity extends Activity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         browserSwitchClient.captureResult(getIntent(), this);
+        if (sReturnIntent != null) {
+            Intent relaunchActivityIntent = new Intent(sReturnIntent)
+                    .addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
+            startActivity(relaunchActivityIntent);
+        }
         finish();
+    }
+
+    public static void setReturnIntent(@NonNull Intent returnIntent) {
+        sReturnIntent = returnIntent;
+    }
+
+    public static void clearReturnIntent() {
+        sReturnIntent = null;
     }
 }