+++ browser-switch/src/main/java/com/braintreepayments/browserswitch/BrowserSwitchConfig.java
@@ -12,7 +12,6 @@ class BrowserSwitchConfig {
 
     Intent createIntentToLaunchUriInBrowser(Context context, Uri uri) {
         Intent intent = new Intent(Intent.ACTION_VIEW, uri);
-        intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
 
         Context applicationContext = context.getApplicationContext();
         if (ChromeCustomTabs.isAvailable(applicationContext)) {