+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # browser-switch-android Release Notes
 
+## TBD
+
+* Use the browser switch with custom tabs in the same activity task as `BraintreeBrowserSwitch`s activity
+* Breaking changes:
+  * The activity which launches the browser switch must now have `android:launchMode="singleTask"`. See the `DemoActivity` entry in the [AndroidManifest.xml](demo/src/main/AndroidManifest.xml) file in the demo application
+  * To ensure that the calling activity will receive the the browser switch result even if the OS kills the app process when the browser is open, you should indicate the activity on app startup. See [DemoApplication](demo/src/main/java/com/braintreepayments/browserswitch/demo/DemoApplication.java) in the demo application:
+    ```java
+    BrowserSwitchActivity.setReturnIntent(new Intent(this, DemoActivity.class))
+    ```
+
 ## 1.1.3
 
 * Update androidx dependencies to latest versions